+++ src/views/auth/signin/index.js
@@ -1,93 +1,19 @@
-import React, { useState } from 'react'
-import { Container, Row, Col, Form } from 'react-bootstrap'
+import React from 'react'
+import { useDispatch } from 'react-redux'
 import { signInAction } from '../../../store/auth/auth.action'
-import { useDispatch, useSelector } from 'react-redux'
-import Loading from '../../../components/loading/index'
-import { SFormSign, SColFooter, STextForm, SButton, STextLink } from '../styled'
+import FormSignIn from '../../../components/portal/auth/signin'
 
-const SignIn = (props) => {
+const SignIn = () => {
   const dispatch = useDispatch()
-  const loading = useSelector((state) => state.auth.loading)
 
-  const [form, setForm] = useState({
-    email: '',
-    password: ''
-  })
-
-  const handleChange = (props) => {
-    const { value, name } = props.target
-    setForm({
-      ...form,
-      [name]: value
-    })
-  }
-
-  const submitForm = (event) => {
-    event.preventDefault()
+  const submitForm = (form) => {
     dispatch(signInAction(form))
   }
 
-  const isNotValid = () => form.email.length === 0 || form.password.length === 0
-
   return (
-    <Container>
-      <Row>
-        <Col md={{ span: 6, offset: 3 }}>
-          <SFormSign>
-            <STextForm>Login</STextForm>
-            <Form.Group className="mb-3">
-              <Form.Label>E-mail</Form.Label>
-              <Form.Control
-                disabled={loading}
-                type="text"
-                name="email"
-                id="email"
-                onChange={handleChange}
-                value={form.email || ''}
-                placeholder="Informe o seu email"
-              />
-            </Form.Group>
-
-            <Form.Group className="mb-3">
-              <Form.Label>Senha</Form.Label>
-              <Form.Control
-                disabled={loading}
-                type="password"
-                name="password"
-                id="password"
-                onChange={handleChange}
-                value={form.password || ''}
-                placeholder="Informe a sua senha"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3">
-              <Form.Check type="checkbox" label="Lembrar credenciais" />
-            </Form.Group>
-
-            {loading ? (
-              <>
-                <Loading />
-              </>
-            ) : (
-              <SButton
-                type="button"
-                disabled={isNotValid()}
-                onClick={submitForm}
-              >
-                {!loading ? 'Entrar' : ''}
-                <i className="icon-angle-right ml-2" />
-              </SButton>
-            )}
-
-            <SColFooter className="text-muted">
-              <h6>Não possui cadastro?</h6>
-              <STextLink href="/registrationclient">Cliente</STextLink>
-              <STextLink href="/registrationprovider">Fornecedor</STextLink>
-            </SColFooter>
-          </SFormSign>
-        </Col>
-      </Row>
-    </Container>
+    <>
+      <FormSignIn submit={submitForm} />
+    </>
   )
 }
 export default SignIn
